Reorganize Exceptions
- Use a dedicated final class to represent asynchronous exceptions. It encapsulates a BalException as a cause, and adds an obligatory correlation identifier. The class should be instantiated only by the blockchain manager because it knows the context + it has the correlation identifier.
