+++ src/main/java/blockchains/iaas/uni/stuttgart/de/adaptation/adapters/ethereum/EthereumAdapter.java
@@ -34,15 +34,15 @@ import blockchains.iaas.uni.stuttgart.de.adaptation.adapters.AbstractAdapter;
 import blockchains.iaas.uni.stuttgart.de.adaptation.utils.BooleanExpressionEvaluator;
 import blockchains.iaas.uni.stuttgart.de.adaptation.utils.PoWConfidenceCalculator;
 import blockchains.iaas.uni.stuttgart.de.adaptation.utils.SmartContractPathParser;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.BalException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.BlockchainNodeUnreachableException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.InvalidScipParameterException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.InvalidTransactionException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.InvokeSmartContractFunctionFailure;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.NotSupportedException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.ParameterException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.SmartContractNotFoundException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.TimeoutException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.BalException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.BlockchainNodeUnreachableException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.InvalidScipParameterException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.InvalidTransactionException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.InvokeSmartContractFunctionFailure;
+import blockchains.iaas.uni.stuttgart.de.exceptions.NotSupportedException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.ParameterException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.SmartContractNotFoundException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.TimeoutException;
 import blockchains.iaas.uni.stuttgart.de.model.Block;
 import blockchains.iaas.uni.stuttgart.de.model.LinearChainTransaction;
 import blockchains.iaas.uni.stuttgart.de.model.Occurrence;
