+++ src/main/java/blockchains/iaas/uni/stuttgart/de/adaptation/adapters/fabric/GatewayManager.java
@@ -20,8 +20,8 @@ import java.util.Map;
 import blockchains.iaas.uni.stuttgart.de.connectionprofiles.AbstractConnectionProfile;
 import blockchains.iaas.uni.stuttgart.de.connectionprofiles.ConnectionProfilesManager;
 import blockchains.iaas.uni.stuttgart.de.connectionprofiles.profiles.FabricConnectionProfile;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.BlockchainNodeUnreachableException;
-import blockchains.iaas.uni.stuttgart.de.scip.model.exceptions.InvalidScipParameterException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.BlockchainNodeUnreachableException;
+import blockchains.iaas.uni.stuttgart.de.exceptions.InvalidScipParameterException;
 import org.hyperledger.fabric.gateway.Contract;
 import org.hyperledger.fabric.gateway.Gateway;
 import org.hyperledger.fabric.gateway.Network;
