+++ src/main/java/blockchains/iaas/uni/stuttgart/de/exceptions/BlockchainIdNotFoundException.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  *******************************************************************************/
-package blockchains.iaas.uni.stuttgart.de.scip.model.exceptions;
+package blockchains.iaas.uni.stuttgart.de.exceptions;
 
 import com.github.arteam.simplejsonrpc.core.annotation.JsonRpcError;
