+++ src/main/java/blockchains/iaas/uni/stuttgart/de/exceptions/BlockchainNodeUnreachableException.java
@@ -8,22 +8,22 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  *******************************************************************************/
-
-package blockchains.iaas.uni.stuttgart.de.scip.model.exceptions;
+package blockchains.iaas.uni.stuttgart.de.exceptions;
 
 import com.github.arteam.simplejsonrpc.core.annotation.JsonRpcError;
 
-@JsonRpcError(code = ExceptionCode.ExecutionError, message = "The execution of the smart contract function resulted in an error.")
-public class InvokeSmartContractFunctionRevoke extends  BalException{
-    public InvokeSmartContractFunctionRevoke() {
+@JsonRpcError(code = ExceptionCode.ConnectionException, message = "The blockchain node cannot be reached.")
+public class BlockchainNodeUnreachableException extends BalException {
+
+    public BlockchainNodeUnreachableException() {
     }
 
-    public InvokeSmartContractFunctionRevoke(String message) {
+    public BlockchainNodeUnreachableException(String message) {
         super(message);
     }
 
     @Override
     public int getCode() {
-        return ExceptionCode.ExecutionError;
+        return ExceptionCode.ConnectionException;
     }
 }
